requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
dfir-unfurl==20201102
dnspython==1.16.0
idna==2.8
ipwhois==1.1.0
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.0
pywin32==224; sys_platform == 'win32'
requests==2.22.0
six==1.12.0
strictyaml==1.2.0
urllib3==1.24.2
win-inet-pton==1.1.0
wincertstore==0.2
wget==3.2